#41 Add logging info when new ip connect to the server
 + Some linting fix


Former-commit-id: 78e8ad13287c3916dc363da0c20276937cd083bd
Former-commit-id: 5e7680c129ca27d9cf0aea32c56f5262aa187103

+++ src/Types.hs
@@ -36,7 +36,7 @@ defaultSendBufferSize :: Int
 defaultSendBufferSize = defaultRecvBufferSize
 
 sO_MARK :: N.SocketOption
-sO_MARK = N.CustomSockOpt (fromIntegral 1, fromIntegral 36) -- https://elixir.bootlin.com/linux/latest/source/arch/alpha/include/uapi/asm/socket.h#L64
+sO_MARK = N.CustomSockOpt (1, 36) -- https://elixir.bootlin.com/linux/latest/source/arch/alpha/include/uapi/asm/socket.h#L64
 
 {-# NOINLINE sO_MARK_Value #-}
 sO_MARK_Value :: IORef Int
